+++ lib/ast-input-transform.js
@@ -2,7 +2,9 @@
 
 const checkAttributes = require('./helpers/check-attributes');
 const supportsAttribute = require('./helpers/supports-attribute');
+const getTag = require('./helpers/get-tag');
 
+const reLines = /(.*?(?:\r\n?|\n|$))/gm;
 const attributeToPropertyMap = {
   role: 'ariaRole',
 };
@@ -85,11 +87,13 @@ const checkboxAttributes = [
 class AngleBracketInputPolyfill {
   constructor(options) {
     this.moduleName = options.meta && options.meta.moduleName;
+    this.syntax = null;
+    this.sourceLines = options.contents && options.contents.match(reLines);
   }
 
   transform(ast) {
     let b = this.syntax.builders;
-    let { moduleName } = this;
+    let { moduleName, sourceLines } = this;
 
     // in order to debug in https://https://astexplorer.net/#/gist/0590eb883edfcd163b183514df4cc717
     // **** copy from here ****
@@ -104,15 +108,15 @@ class AngleBracketInputPolyfill {
 
     let visitor = {
       ElementNode(node) {
-        let tag = node.tag.toLowerCase();
+        let tag = getTag(node, sourceLines);
 
-        if (tag === 'input' || tag === 'textarea') {
+        if (tag === 'Input' || tag === 'Textarea') {
           let { attributes } = node;
           let isCheckbox =
-            tag === 'input' &&
+            tag === 'Input' &&
             attributes.find(({ name, value }) => name === '@type' && value.chars === 'checkbox');
           let supportedAttributes =
-            tag === 'textarea'
+            tag === 'Textarea'
               ? textareaAttributes
               : isCheckbox
               ? checkboxAttributes
@@ -137,7 +141,7 @@ class AngleBracketInputPolyfill {
             )
           );
 
-          return b.mustache(b.path(tag), null, hash, false, node.loc);
+          return b.mustache(b.path(tag.toLowerCase()), null, hash, false, node.loc);
         }
       },
     };
