+++ lib/ast-transform.js
@@ -1,8 +1,10 @@
 'use strict';
 
+const getTag = require('./helpers/get-tag');
+const sourceForNode = require('./helpers/source-for-node');
+
 const reLines = /(.*?(?:\r\n?|\n|$))/gm;
 const ALPHA = /[A-Za-z]/;
-const WHITESPACE = /[\t\r\n\f ]/;
 
 class AngleBracketPolyfill {
   constructor(options) {
@@ -50,37 +52,6 @@ class AngleBracketPolyfill {
       }
     }
 
-    // politely lifted from https://github.com/glimmerjs/glimmer-vm/blob/v0.35.0/packages/%40glimmer/syntax/lib/parser.ts#L113-L149
-    function sourceForNode(node) {
-      let firstLine = node.loc.start.line - 1;
-      let currentLine = firstLine - 1;
-      let firstColumn = node.loc.start.column;
-      let string = [];
-      let line;
-
-      let lastLine = node.loc.end.line - 1;
-      let lastColumn = node.loc.end.column;
-
-      while (currentLine < lastLine) {
-        currentLine++;
-        line = sourceLines[currentLine];
-
-        if (currentLine === firstLine) {
-          if (firstLine === lastLine) {
-            string.push(line.slice(firstColumn, lastColumn));
-          } else {
-            string.push(line.slice(firstColumn));
-          }
-        } else if (currentLine === lastLine) {
-          string.push(line.slice(0, lastColumn));
-        } else {
-          string.push(line);
-        }
-      }
-
-      return string.join('\n');
-    }
-
     function getSelfClosing(element) {
       if ('selfClosing' in element) {
         return element.selfClosing;
@@ -89,51 +60,14 @@ class AngleBracketPolyfill {
         return false;
       }
 
-      let nodeSource = sourceForNode(element);
+      let nodeSource = sourceForNode(element, sourceLines);
       let firstClosingBracketIndex = nodeSource.indexOf('>');
 
       return nodeSource[firstClosingBracketIndex - 1] === '/';
     }
 
-    /*
-      Ember < 3.1.0 mutates `element.tag` so that the first character
-      is _always_ lower case (even if the raw source includes upper case).
-
-      This function falls back to the raw source when possible to detect
-      the _actual_ first char.
-    */
-    function getTag(element) {
-      // if we have no source, we must use whatever element.tag has
-      if (!hasSourceAvailable) {
-        return element.tag;
-      }
-
-      let nodeSource = sourceForNode(element);
-
-      let tagNameStarted = false;
-      let tagName = '';
-      for (let i = 1 /* starting after the opening < */; i < nodeSource.length; i++) {
-        let char = nodeSource[i];
-
-        if (tagNameStarted) {
-          if (char === '/' || char === '>' || WHITESPACE.test(char)) {
-            break;
-          } else {
-            tagName += char;
-          }
-        } else {
-          if (char == '@' || ALPHA.test(char)) {
-            tagNameStarted = true;
-            tagName += char;
-          }
-        }
-      }
-
-      return tagName;
-    }
-
     function getInvocationDetails(element) {
-      let tag = getTag(element);
+      let tag = getTag(element, sourceLines);
       let invocationFirstChar = tag[0];
       let isNamedArgument = invocationFirstChar === '@';
       let isThisPath = tag.indexOf('this.') === 0;
